Add adjustable composite tile layer
This commit adds a an "Adjustable composite" link below all the stacks of a
project. This link allows to create on-the-fly tiles based ot the intensity
settings in the GUI. There, sliders allow to adjust the intensity of each
stack in the project. Dapending on the stack(s) this could involve quite
some memory and CPU requirements. The server configuration might be
tweaked to cope with that.